requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
blas=1.0=mkl
bottleneck=1.3.4=py37hce1f21e_0
brotli=1.0.9=he6710b0_2
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.07.19=h06a4308_0
cairo=1.16.0=hf32fb01_1
certifi=2022.9.24=py37h06a4308_0
colorama=0.4.4=pyhd3eb1b0_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
expat=2.4.4=h295c915_0
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freeglut=3.0.0=hf484d3e_5
freetype=2.11.0=h70c0345_0
fvcore=0.1.5.post20210915=py37
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
graphite2=1.3.14=h23475e2_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=2.8.0=py37h989c5e5_3
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=he6710b0_3
intel-openmp=2021.4.0=h06a4308_3561
iopath=0.1.9=py37
jasper=2.0.14=hd8c5072_2
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h7f8727e_0
kiwisolver=1.3.2=py37h295c915_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libglu=9.0.0=hf484d3e_1
libgomp=9.3.0=h5101ec6_17
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h295c915_1
matplotlib=3.5.1=py37h06a4308_1
matplotlib-base=3.5.1=py37ha18d171_1
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
munkres=1.1.4=py_0
ncurses=6.3=h7f8727e_2
numexpr=2.8.1=py37h6abb31d_0
numpy=1.21.5=py37he7a7128_1
numpy-base=1.21.5=py37hf524024_1
opencv=3.4.2=py37h6fd60c2_1
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
palettable=3.3.0=pypi_0
pandas=1.3.5=py37h8c16a72_0
pcre=8.45=h295c915_0
pillow=9.0.1=py37h22f2fdc_0
pip=21.2.2=py37h06a4308_0
pixman=0.40.0=h7f8727e_1
portalocker=2.5.0=py37h89c1867_0
py-opencv=3.4.2=py37hb342d67_1
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py37h05f1152_2
python=3.7.13=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.7=2_cp37m
pytorch=1.11.0=py3.7_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2022.1=py37h06a4308_0
pyyaml=6.0=py37h7f8727e_1
qt=5.9.7=h5867ecd_1
readline=8.1.2=h7f8727e_1
scikit-learn=1.0.2=py37h51133e4_1
scipy=1.7.3=py37hc147768_0
seaborn=0.11.2=pyhd3eb1b0_0
setuptools=61.2.0=py37h06a4308_0
sip=4.19.8=py37hf484d3e_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.2=hc218d9a_0
tabulate=0.8.10=pyhd8ed1ab_0
termcolor=1.1.0=pyhd8ed1ab_3
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h1ccaba5_0
tornado=6.1=py37h27cfd23_0
tqdm=4.63.0=pyhd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yacs=0.1.6=pyhd3eb1b0_1
yaml=0.2.5=h7b6447c_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
